[test] move IO binding test to onnxruntime_provider_tests (#27055)
### Description

move IO binding test to onnxruntime_provider_tests from
onnxruntime_test_all

### Motivation and Context

Currently, a few unit test cases are included in `onnxruntime_test_all`.
This works well until we want to support WebGPU EP as a plugin EP.

The main differences between `onnxruntime_test_all` and
`onnxruntime_provider_tests` is that the latter respects plugin EP
configuration. The IO Binding test cases involve EP configuration (eg.
"enableGraphCapture" for WebGPU) so it no longer works with WebGPU EP.
